# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для анализа композиционного центра произведений искусства",
        "title": "Art Analysis API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/analysis_order/current": {
            "get": {
                "description": "Возвращает ID текущего чернового заказа и количество добавленных экспертов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AnalysisOrders"
                ],
                "summary": "Получить информацию о текущем черновом заказе",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_CurrTaskInfo"
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/draft/experts/{id}": {
            "post": {
                "description": "Добавляет указанного эксперта в текущий черновой заказ пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Добавить эксперта в черновик заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID эксперта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_OrderExpertLink"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при добавлении эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/experts": {
            "get": {
                "description": "Возвращает список всех экспертов. Поддерживает фильтрацию по названию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Получить список экспертов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию эксперта (подстрока)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.DTO_Resp_Expert"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт нового эксперта с указанными атрибутами.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Создать нового эксперта",
                "parameters": [
                    {
                        "description": "Данные нового эксперта",
                        "name": "expert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_ExpertCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ArtExpert"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/experts/{id}": {
            "get": {
                "description": "Возвращает полную информацию об эксперте по его идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Получить эксперта по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID эксперта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ArtExpert"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Эксперт не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет параметры существующего эксперта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Обновить эксперта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID эксперта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные эксперта",
                        "name": "expert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Req_ExpertCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ArtExpert"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные или ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при обновлении эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Полностью удаляет эксперта по ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Удалить эксперта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID эксперта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_SimpleID"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении эксперта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/experts/{id}/image": {
            "post": {
                "description": "Загружает изображение и сохраняет URL в базе данных.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experts"
                ],
                "summary": "Загрузить изображение эксперта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID эксперта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение эксперта",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.DTO_Resp_UploadImg"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID или отсутствует файл",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при загрузке изображения",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.DTO_Req_ExpertCreate": {
            "description": "Данные для добавления нового эксперта",
            "type": "object",
            "required": [
                "algorithm",
                "name",
                "title"
            ],
            "properties": {
                "algorithm": {
                    "type": "string",
                    "example": "luminosity_center"
                },
                "description": {
                    "type": "string",
                    "example": "Использует метод яркостного центра"
                },
                "img_url": {
                    "type": "string",
                    "example": "expert_photo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов И.И."
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Центр композиционного анализа №1"
                }
            }
        },
...
        "handler.DTO_Resp_CurrTaskInfo": {
            "description": "Статистика по текущей задаче пользователя",
            "type": "object",
            "properties": {
                "experts_count": {
                    "type": "integer",
                    "example": 3
                },
                "order_id": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handler.DTO_Resp_Expert": {
            "description": "Полная информация об эксперте",
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string",
                    "example": "geometric_center"
                },
                "description": {
                    "type": "string",
                    "example": "Определяет композиционный центр"
                },
                "id_artcenter": {
                    "type": "integer",
                    "example": 1
                },
                "img_url": {
                    "type": "string",
                    "example": "expert_photo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов И.И."
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Центр анализа №1"
                }
            }
        },
...
        "handler.DTO_Resp_OrderExpertLink": {
            "description": "Информация о связи между задачей и гейтом",
            "type": "object",
            "properties": {
                "id_artcenter": {
                    "type": "integer",
                    "example": 3
                },
                "id_order": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.DTO_Resp_SimpleID": {
            "description": "Ответ, содержащий только идентификатор",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handler.DTO_Resp_UploadImg": {
            "description": "Результат загрузки изображения эксперта",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "expert_image.png"
                }
            }
        },
...
        "model.AnalysisOrder": {
            "type": "object",
            "properties": {
                "dateCompleted": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "dateFormed": {
                    "type": "string"
                },
                "expertsLinks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExpertsToOrders"
                    }
                },
                "id_creator": {
                    "type": "integer"
                },
                "id_moderator": {
                    "type": "integer"
                },
                "id_order": {
                    "type": "integer"
                },
                "moderator": {
                    "$ref": "#/definitions/model.Users"
                },
                "orderStatus": {
                    "type": "string"
                },
                "resultX": {
                    "type": "number"
                },
                "resultY": {
                    "type": "number"
                },
                "user": {
                    "$ref": "#/definitions/model.Users"
                }
            }
        },
        "model.ArtExpert": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id_artcenter": {
                    "type": "integer"
                },
                "imgURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExpertsToOrders"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ExpertsToOrders": {
            "type": "object",
            "properties": {
                "analysisOrder": {
                    "$ref": "#/definitions/model.AnalysisOrder"
                },
                "artExpert": {
                    "$ref": "#/definitions/model.ArtExpert"
                },
                "centerX": {
                    "type": "number",
                    "format": "float32"
                },
                "centerY": {
                    "type": "number",
                    "format": "float32"
                },
                "id_artcenter": {
                    "type": "integer"
                },
                "id_order": {
                    "type": "integer"
                }
            }
        },
        "model.Users": {
            "type": "object",
            "properties": {
                "id_user": {
                    "type": "integer"
                },
                "isModerator": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}